requirements.txt
asn1crypto==0.24.0
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
DateTime==4.3
future==0.17.1
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
gspread==3.1.0
httplib2==0.13.1
idna==2.8
oauth2client==4.1.3
oauthlib==3.1.0
pprintpp==0.4.0
pyasn1==0.4.6
pyasn1-modules==0.2.6
pybin==0.3.4
pycparser==2.19
pyTelegramBotAPI==3.6.6
python-telegram-bot==12.0.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
telegram==0.0.1
tornado==6.0.3
uritemplate==3.0.0
urllib3==1.25.3
virtualenv==16.7.5
zope.interface==4.6.0
